Add E+ dump clustering, CSV flood AFDD sim, and retire vibe19 parity scripts.
Rename user-facing tooling to eplus dump/clustering paths while keeping
API shims. Synthetic-59 soaks are OpenFDD-only; docs and MCP/agent spec
updated. Vibe19 dual-parity scripts moved to scripts/retired/vibe19-parity/.
