+++ docs/drivers/csv-batch.md
@@ -27,6 +27,8 @@ Execute requires preflight `verdict: pass` (or `warn` when strict mode allows).
 
 Hourly append after a seeded package: `POST /api/csv/import/package/append` (JWT, `confirm: true`). See [CSV batch import](../web-app/csv-batch-import.html).
 
+**AFDD routine sim (BUILDING_50):** [`scripts/csv_flood_afdd_routine_sim.py`](../scripts/csv_flood_afdd_routine_sim.py) — seed + hourly append + `PUT /api/fdd/session-config` patches + `POST /api/fdd/run`. Doc: [CSV flood + AFDD routine](../agent/CSV_FLOOD_AFDD_ROUTINE.md).
+
 ## Host batch (daily pull example)
 
 ```bash
