+++ mcp/README.md
@@ -119,7 +119,7 @@ docker run -i --rm --network host \
 | `openfdd_fdd_equipment` | Loaded equipment IDs/types |
 | `openfdd_fdd_results` | Latest per-equipment rule outcomes |
 | `openfdd_fdd_series` | Mapped live display series |
-| `openfdd_fdd_session_config` | Active units / role map / tuning |
+| `openfdd_fdd_session_config` | GET/PUT units, role map, rule params (AFDD routine tuning) |
 | `openfdd_fdd_accuracy_snapshot` | MCP vs central count parity check |
 | `openfdd_fdd_rule_test_sql` | Test rule SQL |
 | `openfdd_fdd_run` | Run typed registry FDD rules (write gate) |
@@ -127,6 +127,8 @@ docker run -i --rm --network host \
 | `openfdd_reports_draft` / `patch` / `render_pdf` | Report → PDF pipeline (write gate) |
 
 Contract: [ingest contract (archive)](../docs/archive/agent/ingest-contract-v1.md) · [MCP docs](https://bbartling.github.io/open-fdd/mcp-agents/mcp.html)
+
+**Host scripts (not MCP):** CSV flood + AFDD routine — [`docs/agent/CSV_FLOOD_AFDD_ROUTINE.md`](../docs/agent/CSV_FLOOD_AFDD_ROUTINE.md). E+ dump/clustering — [`docs/agent/EPLUS_DUMP_CLUSTERING.md`](../docs/agent/EPLUS_DUMP_CLUSTERING.md).
 
 Bench agent prompt (archive): [docs/archive/agent/bench-driver-setup-wsl-agent.md](../docs/archive/agent/bench-driver-setup-wsl-agent.md)
