+++ openfdd_agent_spec/skills/openfdd-package-mapping/SKILL.md
@@ -26,7 +26,7 @@ and [`docs/migration/vibe19/ROLE_MAPPING_PARITY.md`](../../../docs/migration/vib
 4. Motor ≠ compressor ≠ valve. Status/cmd before amps. Never CHW pump or `clg_valve_pct` as compressor proof. Never motor hours from leave temp.
 5. Weather lives at `{building}/weather/` with `web-outside-air-temp` → `web_oa_t`. Lat/lon in preprocess. `prefer_web_oat`.
 6. `timestamp_utc` is RFC3339 (`Z` or `+00:00`). String `"equip": "AHU_1"` is metadata.
-7. Import with existing `POST /api/csv/import/package` + MCP `openfdd_csv_import_*` / `openfdd_csv_package_append`. Do **not** implement SCAFFOLD `mapping_suggest` / `package_preflight` this cycle.
+7. Import with existing `POST /api/csv/import/package` + MCP `openfdd_csv_import_*` / `openfdd_csv_package_append`. For **streaming sim** on Liberty B50: host script `scripts/csv_flood_afdd_routine_sim.py` (append + AFDD routine patches). Do **not** implement SCAFFOLD `mapping_suggest` / `package_preflight` this cycle.
 8. No vendor/city hardcoding in product. Gold ids: `AHU_1` / `VAV_1` / `CHW_1` / `weather/`.
 
 ## Anti-patterns
