+++ scripts/retired/vibe19-parity/materialize_b100_sql_parity_fixtures.py
