+++ scripts/retired/vibe19-parity/materialize_vav_oracle_fixtures.py
