+++ scripts/synthetic_59_health_matrix_fault_hours_soak.py
@@ -19,12 +19,11 @@ import urllib.request
 from datetime import datetime, timezone
 from pathlib import Path
 
+from eplus_paths import synthetic_artifacts_dir, synthetic_fixture_dir
+
 ROOT = Path(__file__).resolve().parents[1]
-FIXTURE = (
-    ROOT
-    / "reports/wattlab-parity/fixtures/synthetic_59/openfdd_synthetic_59_rule_fixture_v1"
-)
-ARTIFACTS = ROOT / "reports/wattlab-parity/artifacts/synthetic_59"
+FIXTURE = synthetic_fixture_dir()
+ARTIFACTS = synthetic_artifacts_dir()
 EXPECTED = FIXTURE / "expected_faults.csv"
 BUILDING_ID = "OPENFDD_SYNTHETIC_59_RULE_WEEK_V1"
 HOURS_TOL = 0.05
