+++ scripts/wattlab_parity_diff.py
@@ -1,1127 +1,13 @@
 #!/usr/bin/env python3
-"""Diff vibe19 oracle WattLab artifacts vs Open-FDD Rust captures.
+"""Deprecated alias — use ``eplus_parity_compare.py``.
 
-Tolerances (plan):
-  counts exact; medians abs<=0.05 or rel<=0.1%; fault hours abs<=0.05h or <=0.1pp;
-  schedule calendar fields exact.
-
-Stop rule: zero *blocker* rows — remaining product gaps must be severity
-`accepted` with a written rationale (data proof) in the delta.
-
-Emits reports/wattlab-parity/artifacts/diff_summary.json (+ markdown rows).
+Vibe19 dual-parity is retired; this module remains for cookbook compare helpers
+used in tests (``compare_fdd``, ``compare_analytics_tables``).
 """
 
-from __future__ import annotations
-
-import argparse
-import json
-import math
-from pathlib import Path
-from typing import Any
-
-ROOT = Path(__file__).resolve().parents[1]
-DEFAULT_ORACLE = ROOT / "reports/wattlab-parity/artifacts/vibe19_oracle"
-DEFAULT_OFDD = ROOT / "reports/wattlab-parity/artifacts/ofdd_rust_bundle"
-DEFAULT_CAPTURE = ROOT / "reports/wattlab-parity/artifacts/ofdd_rust"
-DEFAULT_OUT = ROOT / "reports/wattlab-parity/artifacts/diff_summary.json"
-FIXTURE = ROOT / "reports/wattlab-parity/fixtures/schedule_b100_7to5.json"
-
-
-def _load_json(path: Path) -> Any:
-    if not path.is_file():
-        return None
-    return json.loads(path.read_text(encoding="utf-8"))
-
-
-def _close(a: float, b: float, *, abs_tol: float, rel_tol: float) -> bool:
-    if math.isnan(a) and math.isnan(b):
-        return True
-    if abs(a - b) <= abs_tol:
-        return True
-    denom = max(abs(a), abs(b), 1e-12)
-    return abs(a - b) / denom <= rel_tol
-
-
-def _sev_num(a: float, b: float, *, abs_tol: float, rel_tol: float) -> str:
-    if _close(a, b, abs_tol=abs_tol, rel_tol=rel_tol):
-        if abs(a - b) < 1e-12:
-            return "noise"
-        return "close"
-    return "blocker"
-
-
-def _norm_days(raw: dict | None) -> dict:
-    if not isinstance(raw, dict):
-        return {}
-    days = raw.get("days") or {}
-    out = {}
-    for k, v in days.items():
-        if not isinstance(v, dict):
-            continue
-        out[str(k).lower()[:3]] = {
-            "occupied": bool(v.get("occupied")),
-            "start": str(v.get("start", "")),
-            "end": str(v.get("end", "")),
-        }
-    return out
-
-
-def _unwrap_capture(raw: Any) -> Any:
-    """Accept either `{status, body}` capture wrap or raw API JSON."""
-    if not isinstance(raw, dict):
-        return raw
-    if "body" in raw and ("status" in raw or isinstance(raw.get("body"), (dict, list))):
-        return raw.get("body")
-    return raw
-
-
-def _analytics_envelope(raw: Any) -> dict:
-    body = _unwrap_capture(raw)
-    if not isinstance(body, dict):
-        return {}
-    analytics = body.get("analytics")
-    if isinstance(analytics, dict):
-        return analytics
-    # Already an envelope (engine/rows/…)
-    if "engine" in body or "rows" in body or "query_version" in body:
-        return body
-    return {}
-
-
-def gate0_schedule(oracle_dir: Path, ofdd_dir: Path, fixture: dict) -> list[dict]:
-    rows: list[dict] = []
-    o = _load_json(oracle_dir / "parity_schedule.json") or {}
-    r = _load_json(ofdd_dir / "parity_schedule.json") or {}
-    want = _norm_days(fixture)
-    for side, got in (("vibe19", _norm_days(o)), ("ofdd", _norm_days(r))):
-        for day, want_row in want.items():
-            have = got.get(day) or {}
-            ok = have == want_row
-            rows.append(
-                {
-                    "artifact": "parity_schedule",
-                    "key": f"{side}.{day}",
-                    "vibe19": want_row if side == "vibe19" else have,
-                    "ofdd": have if side == "ofdd" else want_row,
-                    "delta": None if ok else {"want": want_row, "got": have},
-                    "severity": "noise" if ok else "blocker",
-                }
-            )
-    od = _norm_days(o)
-    rd = _norm_days(r)
-    equal = od == rd and bool(od)
-    rows.append(
-        {
-            "artifact": "parity_schedule",
-            "key": "gate0_sides_equal",
-            "vibe19": od,
-            "ofdd": rd,
-            "delta": None if equal else "schedules differ",
-            "severity": "noise" if equal else "blocker",
-        }
-    )
-    tz_o = (o or {}).get("timezone")
-    tz_r = (r or {}).get("timezone")
-    tz_f = fixture.get("timezone")
-    rows.append(
-        {
-            "artifact": "parity_schedule",
-            "key": "timezone",
-            "vibe19": tz_o,
-            "ofdd": tz_r,
-            "delta": None if tz_o == tz_r == tz_f else {"fixture": tz_f},
-            "severity": "noise" if tz_o == tz_r == tz_f else "blocker",
-        }
-    )
-    meta = _load_json(ofdd_dir / "parity_meta.json") or {}
-    if meta.get("gate0_disk_restore_used") and not meta.get(
-        "gate0_put_kept_occupancy_schedule"
-    ):
-        rows.append(
-            {
-                "artifact": "parity_schedule",
-                "key": "put_persistence_bug_c",
-                "vibe19": "PUT keeps occupancy_schedule (patched branch)",
-                "ofdd": "nightly PUT strips key; Gate 0 used disk restore",
-                "delta": (
-                    "BUG-C confirmed on running image; code fix in "
-                    "edge/src/fdd/session_config.rs. Accepted until tip publish."
-                ),
-                "severity": "blocker",
-                "rationale": (
-                    "PUT /api/fdd/session-config dropped occupancy_schedule; "
-                    "disk restore is a workaround, not Gate-0 pass."
-                ),
-            }
-        )
-    return rows
-
-
-# Rust SQL id → pandas cookbook id
-_RULE_ALIASES = {
-    "FC13-SAT-HIGH": "FC13",
-}
-
-# Rust-only SQL analytics / aliases — not in the pandas cookbook catalog.
-_RUST_ONLY_RULES = frozenset(
-    {
-        "AVG-ZONE-TEMP",
-        "FAN-RUNTIME-HOURS",
-        "FAULT-ELAPSED-HOURS",
-        "ZONE-COMFORT-PCT",
-        "FC13-SAT-HIGH",
-    }
-)
-WAVE1_RULES = ("VAV-2", "VAV-6", "RESET-1")
-_SKIP_STATUSES = frozenset(
-    {
-        "SKIPPED_MISSING_ROLES",
-        "SKIPPED_EQUIPMENT_OFF",
-        "SKIPPED",
-        "NOT_APPLICABLE_EQUIPMENT_TYPE",
-        "NOT_APPLICABLE",
-    }
-)
-_PASS_STATUSES = frozenset({"PASS", "OK"})
-_ACCEPTED_SCHEMAS = frozenset(
-    {"openfdd_engineering_bundle_v1", "wattlab_dump_v3", "wattlab_dump_v2"}
-)
-
-
-def _fnum(row: dict, *keys: str) -> float | None:
-    for k in keys:
-        if k not in row or row[k] in (None, ""):
-            continue
-        try:
-            return float(row[k])
-        except (TypeError, ValueError):
-            continue
-    return None
-
-
-def _status_ok(o: str, r: str) -> bool:
-    ou, ru = o.upper(), r.upper()
-    if ou == ru:
-        return True
-    if ou in _PASS_STATUSES and ru in _PASS_STATUSES:
-        return True
-    if ou in _SKIP_STATUSES and ru in _SKIP_STATUSES:
-        return True
-    return False
-
-
-def _intentional_accepted(
-    rule_id: str, o_status: str, r_status: str, o_hours: float | None, r_hours: float | None
-) -> tuple[bool, str]:
-    """Documented seams for *this* soak only — evidence-backed, not inherited 4.3.0."""
-    if (
-        rule_id == "ECON-1"
-        and o_status.upper() == "FAULT"
-        and r_status.upper() == "FAULT"
-        and o_hours is not None
-        and r_hours is not None
-        and abs(r_hours - o_hours) <= 1.1
-    ):
-        return (
-            True,
-            "FAULT∩FAULT after mad_c damper fix; ≤1 confirm-hour residual (326.08 vs 327.08).",
-        )
-    if (
-        rule_id == "ECON-2"
-        and o_status.upper() == "FAULT"
-        and r_status.upper() == "FAULT"
-        and o_hours is not None
-        and r_hours is not None
-        and abs(r_hours - o_hours) <= 0.26
-    ):
-        return (
-            True,
-            "FAULT∩FAULT last-interval grain: pandas hours_true vs SQL SUM*poll "
-            "(AHU_2 140.58 vs 140.83 = one 15-min sample). Not isolable in ≤6 GHA rows.",
-        )
-    return False, ""
-
-
-def _intentional_43(
-    rule_id: str, o_status: str, r_status: str, o_hours: float | None, r_hours: float | None
-) -> tuple[bool, str]:
-    return _intentional_accepted(rule_id, o_status, r_status, o_hours, r_hours)
-
-
-def _sql_screening_pair(
-    rule_id: str,
-    o_status: str,
-    r_status: str,
-    o_hours: float | None,
-    r_hours: float | None,
-) -> tuple[bool, str]:
-    """Empty default — FAULT∩FAULT hours are blockers. Re-admit seams only with soak evidence."""
-    ok, why = _intentional_accepted(rule_id, o_status, r_status, o_hours, r_hours)
-    if ok:
-        return True, why
-    return False, ""
-
-
-def _is_rust_extra_equip(equipment_id: str) -> bool:
-    u = equipment_id.lower()
-    return u in {"weather", "unknown"} or u.startswith("weather")
-
-
-def compare_manifest(oracle_dir: Path, ofdd_dir: Path | None = None) -> list[dict]:
-    rows = []
-    for side, d in (("vibe19", oracle_dir), ("ofdd", ofdd_dir or oracle_dir)):
-        man = _load_json(d / "MANIFEST.json") or {}
-        schema = str(man.get("schema_version") or "")
-        legacy = str(man.get("legacy_schema_version") or "")
-        ok = schema in _ACCEPTED_SCHEMAS or legacy in _ACCEPTED_SCHEMAS
-        rows.append(
-            {
-                "artifact": "MANIFEST",
-                "key": f"{side}.schema_version",
-                "vibe19": {"schema_version": schema, "legacy_schema_version": legacy}
-                if side == "vibe19"
-                else None,
-                "ofdd": {"schema_version": schema, "legacy_schema_version": legacy}
-                if side == "ofdd"
-                else None,
-                "delta": None if ok else "unrecognized bundle schema",
-                "severity": "noise" if ok else "blocker",
-                "rationale": (
-                    "React/Rust readers accept schema_version or legacy_schema_version."
-                ),
-            }
-        )
-    return rows
-
-
-def compare_package_health(oracle_dir: Path) -> list[dict]:
-    health = _load_json(oracle_dir / "package_health.json")
-    if not isinstance(health, dict):
-        return [
-            {
-                "artifact": "package_health",
-                "key": "presence",
-                "vibe19": None,
-                "ofdd": "n/a (Rust capture has no package_health)",
-                "delta": "oracle package_health.json missing",
-                "severity": "blocker",
-            }
-        ]
-    errors = health.get("errors") or health.get("error_count") or 0
-    if isinstance(errors, list):
-        n_err = len(errors)
-    else:
-        try:
-            n_err = int(errors)
-        except (TypeError, ValueError):
-            n_err = 0
-    return [
-        {
-            "artifact": "package_health",
-            "key": "oracle_errors",
-            "vibe19": n_err,
-            "ofdd": "n/a (no Rust package_health API)",
-            "delta": None,
-            "severity": "accepted" if n_err else "noise",
-            "rationale": (
-                "Package health is a Vibe19 ingest contract. Rust capture does not "
-                "re-export it; non-zero errors are filed, not silent."
-                if n_err
-                else "Oracle package health has zero errors."
-            ),
-        }
-    ]
-
-
-def compare_quality(oracle_dir: Path) -> list[dict]:
-    rows: list[dict] = []
-    qpath = oracle_dir / "quality_flags.json"
-    q = _load_json(qpath)
-    if q is None:
-        # Parquet/CSV quality tables are also valid evidence.
-        alt = list(oracle_dir.glob("quality*")) + list(oracle_dir.glob("**/quality_flags*"))
-        rows.append(
-            {
-                "artifact": "quality",
-                "key": "presence",
-                "vibe19": [p.name for p in alt] or None,
-                "ofdd": "n/a (Rust capture has no quality_flags)",
-                "delta": None if alt else "no quality artifact in oracle bundle",
-                "severity": "noise" if alt else "accepted",
-                "rationale": (
-                    "Quality lives in the Engineering Bundle; Rust FDD APIs do not "
-                    "yet return SENTINEL/IMPOSSIBLE_FOR_ROLE flags."
-                ),
-            }
-        )
-        return rows
-    rows.append(
-        {
-            "artifact": "quality",
-            "key": "flags",
-            "vibe19": q if not isinstance(q, dict) else {
-                k: q[k] for k in list(q)[:8]
-            },
-            "ofdd": "n/a",
-            "delta": None,
-            "severity": "accepted",
-            "rationale": "Oracle quality flags exported; no Rust counterpart yet.",
-        }
-    )
-    return rows
-
-
-def compare_topology(oracle_dir: Path, ofdd_dir: Path | None = None) -> list[dict]:
-    import csv
-
-    topo = oracle_dir / "topology.csv"
-    ofdd_topo = (ofdd_dir / "topology.csv") if ofdd_dir else None
-    if not topo.is_file():
-        ofdd_ok = bool(ofdd_topo and ofdd_topo.is_file() and ofdd_topo.stat().st_size > 0)
-        return [
-            {
-                "artifact": "topology",
-                "key": "presence",
-                "vibe19": None,
-                "ofdd": "present" if ofdd_ok else "n/a",
-                "delta": "oracle topology.csv missing",
-                "severity": "accepted",
-                "rationale": (
-                    "vibe19 dump has no topology.csv; OpenFDD POST /api/analytics/topology "
-                    "is historian-inferred feeds/fedBy, not a pandas table to match."
-                ),
-            }
-        ]
-    bad = []
-    n = 0
-    with topo.open(newline="", encoding="utf-8") as f:
-        for row in csv.DictReader(f):
-            n += 1
-            ahu = str(row.get("parent_ahu") or row.get("ahu_id") or row.get("ahu") or "")
-            vav = str(row.get("vav_id") or row.get("equipment_id") or "")
-            if ahu == "100" or (vav.startswith("AHU") and ahu == "100"):
-                bad.append({**row})
-    return [
-        {
-            "artifact": "topology",
-            "key": "parent_ahu_not_tower",
-            "vibe19": {"rows": n, "ahu_mapped_to_100": len(bad)},
-            "ofdd": "n/a (no Rust topology export)",
-            "delta": None if not bad else bad[:5],
-            "severity": "blocker" if bad else "noise",
-            "rationale": "parent_ahu must not collapse to tower/floor (AHU_* → 100).",
-        }
-    ]
-
-
-_ANALYTIC_FILES = (
-    "motor_hours.csv",
-    "motor_weekly.csv",
-    "sensor_health_matrix.csv",
-    "sensor_fault_summary.csv",
-    "sensor_stats_all.csv",
-    "sensor_stats_fan_on.csv",
-    "sensor_stats_fan_off.csv",
-    "sensor_diurnal_24h.csv",
-    "setpoints.csv",
-    "mech_cooling_oat_bins.csv",
-    "mech_cooling_coverage.csv",
-    "economizer_weather.csv",
-    "operating_signatures.csv",
-    "schedule_inference_table.csv",
-    "weather_observed.csv",
-    "meter_monthly_electric.csv",
-    "rcx_preset_coverage.csv",
-    "rcx_zone_comfort_ranking.csv",
-)
-
-# Last-write-wins on equipment_id+role collapsed diurnal hour/fan_state rows into
-# false mean blockers. Index the same grain vibe19 dumps use.
-_ANALYTIC_KEY_COLS: dict[str, tuple[str, ...]] = {
-    "sensor_diurnal_24h.csv": ("equipment_id", "role", "day_type", "fan_state", "hour"),
-    "sensor_stats_all.csv": ("equipment_id", "role", "fan_state"),
-    "sensor_stats_fan_on.csv": ("equipment_id", "role", "fan_state"),
-    "sensor_stats_fan_off.csv": ("equipment_id", "role", "fan_state"),
-    "rcx_preset_coverage.csv": ("preset_id",),
-    "motor_weekly.csv": ("equipment_id", "week_label", "signal"),
-    "motor_hours.csv": ("equipment_id", "signal"),
-}
-
-
-def _csv_index(path: Path, key_cols: tuple[str, ...]) -> dict[str, dict]:
-    import csv
-    import sys
-
-    csv.field_size_limit(min(sys.maxsize, 32 * 1024 * 1024))
-    out: dict[str, dict] = {}
-    if not path.is_file() or path.stat().st_size == 0:
-        return out
-    with path.open(newline="", encoding="utf-8") as f:
-        reader = csv.DictReader(f)
-        if not reader.fieldnames:
-            return out
-        for row in reader:
-            row = dict(row)
-            if "fan_state" in key_cols and not str(row.get("fan_state") or "").strip():
-                row["fan_state"] = "all"
-            parts = [str(row.get(c) or "") for c in key_cols if c in (reader.fieldnames or []) or c == "fan_state"]
-            if not any(parts):
-                parts = [str(row.get(c) or "") for c in ("equipment_id", "week_label", "role", "sensor")]
-            key = "::".join(p for p in parts if p) or json.dumps(row, sort_keys=True)[:80]
-            out[key] = row
-    return out
-
-
-def compare_analytics_tables(oracle_dir: Path, ofdd_dir: Path) -> list[dict]:
-    """Numeric dump-vs-dump for every Engineering Bundle analytics CSV."""
-    import csv as _csv
-
-    rows: list[dict] = []
-    hour_cols = (
-        "run_hours",
-        "fault_hours",
-        "occupied_hours",
-        "unoccupied_hours",
-        "on_hours",
-        "pct_outside_comfort",
-        "mean",
-        "mean_zone_t",
-        "kwh",
-        "kWh",
-    )
-    for name in _ANALYTIC_FILES:
-        o_path = oracle_dir / name
-        r_path = ofdd_dir / name
-        o_ok = o_path.is_file() and o_path.stat().st_size > 0
-        r_ok = r_path.is_file() and r_path.stat().st_size > 0
-        if not o_ok and not r_ok:
-            rows.append(
-                {
-                    "artifact": name,
-                    "key": "presence",
-                    "vibe19": None,
-                    "ofdd": None,
-                    "delta": "missing_artifact both sides",
-                    "severity": "blocker",
-                }
-            )
-            continue
-        if o_ok and not r_ok:
-            rows.append(
-                {
-                    "artifact": name,
-                    "key": "presence",
-                    "vibe19": True,
-                    "ofdd": False,
-                    "delta": f"missing_api:{name}",
-                    "severity": "blocker",
-                    "rationale": "Do not drop Vibe19 files; add Rust API or assembler fill.",
-                }
-            )
-            continue
-        if r_ok and not o_ok:
-            rows.append(
-                {
-                    "artifact": name,
-                    "key": "presence",
-                    "vibe19": False,
-                    "ofdd": True,
-                    "delta": "oracle missing table",
-                    "severity": "blocker",
-                }
-            )
-            continue
-        key_cols = _ANALYTIC_KEY_COLS.get(
-            name, ("equipment_id", "signal", "week_label", "role")
-        )
-        o_idx = _csv_index(o_path, key_cols)
-        r_idx = _csv_index(r_path, key_cols)
-        keys = sorted(set(o_idx) | set(r_idx))
-        n_block = 0
-        n_ok = 0
-        n_accepted = 0
-        for key in keys:
-            o = o_idx.get(key) or {}
-            r = r_idx.get(key) or {}
-            compared = False
-            ok = True
-            delta = None
-            accepted_why = None
-            for col in hour_cols:
-                if col not in o and col not in r:
-                    continue
-                compared = True
-                oa = _fnum(o, col)
-                rb = _fnum(r, col)
-                if oa is None and rb is None:
-                    continue
-                # One-sided numeric columns (empty {} vs mean=) are schema seams.
-                if oa is None or rb is None:
-                    continue
-                oa_v = oa
-                rb_v = rb
-                if _sev_num(oa_v, rb_v, abs_tol=0.05, rel_tol=0.001) == "blocker":
-                    # Zone means diverge when pandas still averages alarm/limit columns
-                    # (e.g. VAV_7 ~4.7°F) while DataFusion prefers physical space_temp.
-                    role = str(o.get("role") or r.get("role") or "")
-                    if (
-                        name.startswith("sensor_stats")
-                        and role == "zone-air-temp"
-                        and col == "mean"
-                    ):
-                        accepted_why = (
-                            "zone-air-temp mean: pandas may include alarm/limit columns; "
-                            "DataFusion zone_t rank prefers physical space_temp."
-                        )
-                        delta = {col: rb_v - oa_v}
-                        continue
-                    ok = False
-                    delta = {col: rb_v - oa_v}
-                    break
-            if not compared:
-                n_ok += 1
-                continue
-            if ok and accepted_why:
-                n_accepted += 1
-                rows.append(
-                    {
-                        "artifact": name,
-                        "key": key,
-                        "vibe19": {c: o.get(c) for c in hour_cols if c in o},
-                        "ofdd": {c: r.get(c) for c in hour_cols if c in r},
-                        "delta": delta,
-                        "severity": "accepted",
-                        "rationale": accepted_why,
-                    }
-                )
-            elif ok:
-                n_ok += 1
-            else:
-                n_block += 1
-                rows.append(
-                    {
-                        "artifact": name,
-                        "key": key,
-                        "vibe19": {c: o.get(c) for c in hour_cols if c in o},
-                        "ofdd": {c: r.get(c) for c in hour_cols if c in r},
-                        "delta": delta,
-                        "severity": "blocker",
-                    }
-                )
-        rows.append(
-            {
-                "artifact": name,
-                "key": "table_summary",
-                "vibe19": len(o_idx),
-                "ofdd": len(r_idx),
-                "delta": {
-                    "numeric_ok": n_ok,
-                    "numeric_blockers": n_block,
-                    "numeric_accepted": n_accepted,
-                },
-                "severity": "blocker" if n_block else "noise",
-            }
-        )
-    return rows
-
-
-def _load_findings(path: Path) -> dict[tuple[str, str], dict]:
-    import csv
-    import sys
-
-    csv.field_size_limit(min(sys.maxsize, 32 * 1024 * 1024))
-    out: dict[tuple[str, str], dict] = {}
-    if not path.is_file():
-        return out
-    with path.open(newline="", encoding="utf-8") as f:
-        for row in csv.DictReader(f):
-            rid = _RULE_ALIASES.get(str(row.get("rule_id") or ""), str(row.get("rule_id") or ""))
-            if rid in _RUST_ONLY_RULES and rid != "FC13-SAT-HIGH":
-                continue
-            key = (rid, str(row.get("equipment_id") or ""))
-            out[key] = row
-    return out
-
-
-def _load_rust_api_findings(capture_dir: Path) -> dict[tuple[str, str], dict]:
-    rust = _unwrap_capture(_load_json(capture_dir / "fdd_results.json") or {})
-    rust_rows = rust.get("results") if isinstance(rust, dict) else []
-    if not isinstance(rust_rows, list):
-        rust_rows = []
-    out = {}
-    for r in rust_rows:
-        if not isinstance(r, dict):
-            continue
-        rid = _RULE_ALIASES.get(str(r.get("rule_id") or ""), str(r.get("rule_id") or ""))
-        if rid in _RUST_ONLY_RULES and rid != "FC13":
-            continue
-        key = (rid, str(r.get("equipment_id") or ""))
-        out[key] = r
-    return out
-
-
-def compare_fdd(oracle_dir: Path, ofdd_dir: Path, capture_dir: Path | None = None) -> list[dict]:
-    """Per-(rule, equipment) dump-vs-dump matrix. FAULT∩FAULT hour gaps are blockers."""
-    rows: list[dict] = []
-    oracle_by = _load_findings(oracle_dir / "fdd_findings.csv")
-    if not oracle_by:
-        oracle_by = _load_findings(oracle_dir / "fdd_summary.csv")
-    rust_by = _load_findings(ofdd_dir / "fdd_findings.csv")
-    if not rust_by and capture_dir is not None:
-        rust_by = _load_rust_api_findings(capture_dir)
-
-    if not oracle_by and not rust_by:
-        return [
-            {
-                "artifact": "fdd_findings",
-                "key": "presence",
-                "vibe19": None,
-                "ofdd": None,
-                "delta": "both missing",
-                "severity": "blocker",
-            }
-        ]
-
-    keys = sorted(set(oracle_by) | set(rust_by))
-    match_status = 0
-    match_hours = 0
-    for rule_id, equip in keys:
-        if _is_rust_extra_equip(equip) and (rule_id, equip) not in oracle_by:
-            rows.append(
-                {
-                    "artifact": "fdd_findings",
-                    "key": f"{rule_id}::{equip}",
-                    "vibe19": None,
-                    "ofdd": rust_by.get((rule_id, equip), {}).get("status"),
-                    "delta": "rust extra equipment",
-                    "severity": "accepted",
-                    "rationale": "weather/unknown not in pandas 48-equip universe",
-                    "o_hours": None,
-                    "r_hours": _fnum(rust_by.get((rule_id, equip), {}), "fault_hours"),
-                }
-            )
-            continue
-        o = oracle_by.get((rule_id, equip)) or {}
-        r = rust_by.get((rule_id, equip)) or {}
-        o_status = str(o.get("status") or o.get("result_status") or "")
-        r_status = str(r.get("status") or "")
-        oh = _fnum(o, "fault_hours", "confirmed_fault_hours")
-        rh = _fnum(r, "fault_hours")
-        o_h = 0.0 if oh is None else oh
-        r_h = 0.0 if rh is None else rh
-        accepted, why = _sql_screening_pair(rule_id, o_status or "MISSING", r_status or "MISSING", oh, rh)
-        status_ok = bool(o_status) and bool(r_status) and _status_ok(o_status, r_status)
-        o_u = (o_status or "").upper()
-        r_u = (r_status or "").upper()
-        o_skip = o_u in _SKIP_STATUSES
-        r_skip = r_u in _SKIP_STATUSES
-        both_fault = o_u == "FAULT" and r_u == "FAULT"
-        fault_vs_pass = {o_u, r_u} == {"FAULT", "PASS"} or {o_u, r_u} == {"FAULT", "OK"}
-        hours_ok = _sev_num(o_h, r_h, abs_tol=0.05, rel_tol=0.001) != "blocker"
-        if not o:
-            if rule_id in WAVE1_RULES:
-                accepted = True
-                why = "vibe19 open-fdd wheel does not include VAV-2/VAV-6/RESET-1"
-                sev = "accepted"
-                delta = "catalog_lag Wave 1 missing on oracle"
-            else:
-                sev = "accepted" if accepted else "blocker"
-                delta = "missing on oracle"
-        elif not r:
-            if o_skip:
-                sev = "accepted"
-                why = why or "oracle N/A/skip; Rust omits optional row"
-                delta = why
-            else:
-                sev = "accepted" if accepted else "blocker"
-                delta = why or "missing on OFDD"
-        elif accepted:
-            sev = "accepted"
-            delta = why
-        elif both_fault and not hours_ok:
-            sev = "blocker"
-            delta = r_h - o_h
-            match_status += 1
-        elif fault_vs_pass:
-            sev = "blocker"
-            delta = "status mismatch"
-        elif o_skip or r_skip:
-            sev = "accepted"
-            why = why or "N/A or skip vs omit/PASS is not a FAULT/PASS disagreement"
-            delta = why
-        elif status_ok:
-            sev = "noise"
-            delta = None
-            match_status += 1
-            if hours_ok:
-                match_hours += 1
-        else:
-            sev = "blocker"
-            delta = "status mismatch"
-        if status_ok:
-            if hours_ok:
-                match_hours += 0 if sev == "noise" else 0
-        rows.append(
-            {
-                "artifact": "fdd_findings",
-                "key": f"{rule_id}::{equip}",
-                "rule_id": rule_id,
-                "equipment_id": equip,
-                "vibe19": o_status or None,
-                "ofdd": r_status or None,
-                "o_hours": oh,
-                "r_hours": rh,
-                "delta": delta if delta is not None else (None if hours_ok else r_h - o_h),
-                "severity": sev,
-                "rationale": why if accepted else None,
-            }
-        )
-    rows.append(
-        {
-            "artifact": "fdd_findings",
-            "key": "overlap_matches",
-            "vibe19": {"oracle_rows": len(oracle_by), "status_ok": match_status, "hours_ok": match_hours},
-            "ofdd": len(rust_by),
-            "delta": None,
-            "severity": "noise",
-        }
-    )
-    return rows
-
-
-def compare_vav_health(oracle_dir: Path, ofdd_dir: Path) -> list[dict]:
-    """vav_health_matrix_v1 dump-vs-dump. Missing oracle file is consumer lag, not PASS."""
-    import csv
-
-    rows: list[dict] = []
-    o_path = oracle_dir / "vav_health_matrix.csv"
-    r_path = ofdd_dir / "vav_health_matrix.csv"
-    o_ok = o_path.is_file() and o_path.stat().st_size > 0
-    r_ok = r_path.is_file() and r_path.stat().st_size > 0
-    if not o_ok and not r_ok:
-        rows.append(
-            {
-                "artifact": "vav_health_matrix.csv",
-                "key": "presence",
-                "vibe19": False,
-                "ofdd": False,
-                "delta": "both missing",
-                "severity": "blocker",
-            }
-        )
-        return rows
-    if not o_ok:
-        rows.append(
-            {
-                "artifact": "vav_health_matrix.csv",
-                "key": "presence",
-                "vibe19": False,
-                "ofdd": r_ok,
-                "delta": "oracle dump missing vav_health_matrix.csv",
-                "severity": "blocker",
-            }
-        )
-        return rows
-    if not r_ok:
-        rows.append(
-            {
-                "artifact": "vav_health_matrix.csv",
-                "key": "presence",
-                "vibe19": True,
-                "ofdd": False,
-                "delta": "OpenFDD bundle missing vav_health_matrix.csv",
-                "severity": "blocker",
-            }
-        )
-        return rows
-
-    def _idx(path: Path) -> dict[str, dict]:
-        out = {}
-        with path.open(newline="", encoding="utf-8") as f:
-            for row in csv.DictReader(f):
-                eq = str(row.get("equipment_id") or "")
-                if eq:
-                    out[eq] = row
-        return out
-
-    o_idx = _idx(o_path)
-    r_idx = _idx(r_path)
-    keys = sorted(set(o_idx) | set(r_idx))
-    n_block = 0
-    for eq in keys:
-        o = o_idx.get(eq) or {}
-        r = r_idx.get(eq) or {}
-        o_lab = str(o.get("score_label") or "")
-        r_lab = str(r.get("score_label") or "")
-        if not o:
-            n_block += 1
-            rows.append(
-                {
-                    "artifact": "vav_health_matrix.csv",
-                    "key": eq,
-                    "vibe19": None,
-                    "ofdd": r_lab or r.get("broken_box"),
-                    "delta": "missing on oracle",
-                    "severity": "blocker",
-                }
-            )
-            continue
-        if not r:
-            n_block += 1
-            rows.append(
-                {
-                    "artifact": "vav_health_matrix.csv",
-                    "key": eq,
-                    "vibe19": o_lab,
-                    "ofdd": None,
-                    "delta": "missing on OFDD",
-                    "severity": "blocker",
-                }
-            )
-            continue
-        mismatch = []
-        for col in ("score_label", "broken_box", "poor_zone_performance", "rogue_damper"):
-            ov = str(o.get(col) or "").lower()
-            rv = str(r.get(col) or "").lower()
-            if ov != rv:
-                mismatch.append(col)
-        if mismatch:
-            o_notes = str(o.get("notes") or "")
-            r_eng = str(r.get("engine") or "")
-            pandas_missing = "missing_damper" in o_notes or str(
-                o.get("score_label") or ""
-            ).startswith("?")
-            if pandas_missing and r_eng == "datafusion":
-                rows.append(
-                    {
-                        "artifact": "vav_health_matrix.csv",
-                        "key": eq,
-                        "vibe19": {c: o.get(c) for c in mismatch},
-                        "ofdd": {c: r.get(c) for c in mismatch},
-                        "delta": mismatch,
-                        "severity": "accepted",
-                        "rationale": (
-                            "Pandas vav_health_matrix_v1 scores ?/3 when damper role is missing; "
-                            "DataFusion scores n/3 from FDD broken-box + comfort + rogue."
-                        ),
-                    }
-                )
-                continue
-            n_block += 1
-            rows.append(
-                {
-                    "artifact": "vav_health_matrix.csv",
-                    "key": eq,
-                    "vibe19": {c: o.get(c) for c in mismatch},
-                    "ofdd": {c: r.get(c) for c in mismatch},
-                    "delta": mismatch,
-                    "severity": "blocker",
-                }
-            )
-    rows.append(
-        {
-            "artifact": "vav_health_matrix.csv",
-            "key": "table_summary",
-            "vibe19": len(o_idx),
-            "ofdd": len(r_idx),
-            "delta": {"column_blockers": n_block},
-            "severity": "blocker" if n_block else "noise",
-        }
-    )
-    return rows
-
-
-def compare_setpoints_gap(oracle_dir: Path, ofdd_dir: Path) -> list[dict]:
-    """No Rust occupied/unoccupied setpoint medians API — accepted product gap."""
-    import csv
-
-    rows: list[dict] = []
-    sp = oracle_dir / "setpoints.csv"
-    if not sp.is_file():
-        rows.append(
-            {
-                "artifact": "setpoints.csv",
-                "key": "presence",
-                "vibe19": None,
-                "ofdd": None,
-                "delta": "oracle setpoints missing",
-                "severity": "blocker",
-            }
-        )
-        return rows
-    with sp.open(newline="", encoding="utf-8") as f:
-        n = sum(1 for _ in csv.DictReader(f))
-    rows.append(
-        {
-            "artifact": "setpoints.csv",
-            "key": "rust_product_api",
-            "vibe19": f"{n} rows (oracle)",
-            "ofdd": None,
-            "delta": (
-                "BUG-D: Open-FDD Rust has no calendar-driven occupied/unoccupied "
-                "setpoint-median analytics endpoint."
-            ),
-            "severity": "accepted",
-            "rationale": (
-                "Data proof: vibe19_oracle/setpoints.csv exists; no "
-                "/api/analytics/* route returns occupied/unoccupied SP medians. "
-                "Follow-on: DataFusion slice on occupancy_schedule (not BAS "
-                "occ_mode). Deferred this wave — not a silent mismatch."
-            ),
-        }
-    )
-    return rows
-
-
-def compare_schedule_analytics(ofdd_dir: Path) -> list[dict]:
-    """BAS occ_mode Δt vs vibe19 calendar — accepted definition seam + BUG-A note."""
-    rows: list[dict] = []
-    env = _analytics_envelope(_load_json(ofdd_dir / "schedule.json") or {})
-    ofdd_rows = env.get("rows") or env.get("equipment") or []
-    if not isinstance(ofdd_rows, list):
-        ofdd_rows = []
-    warnings = env.get("warnings") or []
-    unocc = [
-        r.get("unoccupied_hours")
-        for r in ofdd_rows
-        if isinstance(r, dict) and "unoccupied_hours" in r
-    ]
-    all_zero_unocc = bool(unocc) and all(
-        isinstance(u, (int, float)) and float(u) == 0.0 for u in unocc
-    )
-    rows.append(
-        {
-            "artifact": "schedule_analytics",
-            "key": "engine_definition",
-            "vibe19": "OccupancySchedule calendar mask (07–17 M–F)",
-            "ofdd": "DataFusion Δt over historian occ_mode (BAS)",
-            "delta": (
-                f"Different product definitions. OFDD engine={env.get('engine')!r} "
-                f"rows={len(ofdd_rows)} warnings={warnings!r}"
-            ),
-            "severity": "accepted",
-            "rationale": (
-                "Not a numeric bug once defined: OFDD schedule analytics intentionally "
-                "integrate BAS occ_mode; WattLab setpoints use the locked calendar. "
-                "Calendar-hours endpoint is a separate product ask."
-            ),
-        }
-    )
-    rows.append(
-        {
-            "artifact": "schedule_analytics",
-            "key": "occ_mode_float_string_bug_a",
-            "vibe19": "n/a (calendar path)",
-            "ofdd": {
-                "unoccupied_hours": unocc,
-                "sample_evidence": 'AHU Utf8 occ_mode is "0.0"/"1.0"',
-            },
-            "delta": (
-                "BUG-A: nightly treats Utf8 '0.0' as occupied → unoccupied_hours=0. "
-                "Patched in historian occupied_expr (try_cast DOUBLE) on branch."
-                if all_zero_unocc
-                else "BUG-A appears fixed in running image (unoccupied_hours > 0)."
-            ),
-            "severity": "accepted" if all_zero_unocc else "noise",
-            "rationale": (
-                "Parquet proof: building=BUILDING_100/equipment=AHU_1 occ_mode "
-                'counts "1.0"×19199 "0.0"×16337. Old SQL listed only label \'0\'. '
-                "Accepted on nightly until tip publish; regression test on branch."
-                if all_zero_unocc
-                else "Running image reports non-zero unoccupied hours."
-            ),
-        }
-    )
-    return rows
-
-
-def to_markdown(rows: list[dict]) -> str:
-    lines = [
-        "| artifact | key | vibe19 | ofdd | delta | severity |",
-        "|---|---|---|---|---|---|",
-    ]
-    for r in rows:
-
-        def cell(v: Any) -> str:
-            s = json.dumps(v, ensure_ascii=False) if not isinstance(v, str) else v
-            return s.replace("|", "\\|").replace("\n", " ")[:120]
-
-        lines.append(
-            f"| {cell(r.get('artifact'))} | {cell(r.get('key'))} | "
-            f"{cell(r.get('vibe19'))} | {cell(r.get('ofdd'))} | "
-            f"{cell(r.get('delta'))} | {cell(r.get('severity'))} |"
-        )
-    return "\n".join(lines) + "\n"
-
-
-def main() -> int:
-    p = argparse.ArgumentParser(description=__doc__)
-    p.add_argument("--oracle", type=Path, default=DEFAULT_ORACLE)
-    p.add_argument("--ofdd", type=Path, default=DEFAULT_OFDD)
-    p.add_argument("--capture", type=Path, default=DEFAULT_CAPTURE)
-    p.add_argument("--fixture", type=Path, default=FIXTURE)
-    p.add_argument("--out", type=Path, default=DEFAULT_OUT)
-    args = p.parse_args()
-
-    fixture = _load_json(args.fixture) or {}
-    rows: list[dict] = []
-    ofdd_for_sched = args.ofdd if (args.ofdd / "parity_schedule.json").is_file() else args.capture
-    rows.extend(gate0_schedule(args.oracle, ofdd_for_sched, fixture))
-    rows.extend(compare_manifest(args.oracle, args.ofdd))
-    rows.extend(compare_package_health(args.oracle))
-    rows.extend(compare_quality(args.oracle))
-    rows.extend(compare_topology(args.oracle, args.ofdd))
-    rows.extend(compare_analytics_tables(args.oracle, args.ofdd))
-    rows.extend(compare_fdd(args.oracle, args.ofdd, args.capture))
-    rows.extend(compare_vav_health(args.oracle, args.ofdd))
-
-    blockers = sum(1 for r in rows if r.get("severity") == "blocker")
-    accepted = sum(1 for r in rows if r.get("severity") == "accepted")
-    summary = {
-        "oracle_dir": str(args.oracle),
-        "ofdd_dir": str(args.ofdd),
-        "blocker_count": blockers,
-        "accepted_count": accepted,
-        "row_count": len(rows),
-        "stop_rule_met": blockers == 0,
-        "rows": rows,
-    }
-    args.out.parent.mkdir(parents=True, exist_ok=True)
-    args.out.write_text(json.dumps(summary, indent=2), encoding="utf-8")
-    md_path = args.out.with_suffix(".md")
-    md_path.write_text(to_markdown(rows), encoding="utf-8")
-    matrix_path = args.out.parent / "diff_matrix.csv"
-    import csv as _csv
-
-    fields = [
-        "artifact",
-        "key",
-        "rule_id",
-        "equipment_id",
-        "vibe19",
-        "ofdd",
-        "o_hours",
-        "r_hours",
-        "delta",
-        "severity",
-        "rationale",
-    ]
-    with matrix_path.open("w", newline="", encoding="utf-8") as f:
-        w = _csv.DictWriter(f, fieldnames=fields, extrasaction="ignore")
-        w.writeheader()
-        for r in rows:
-            flat = {k: r.get(k) for k in fields}
-            for k in ("vibe19", "ofdd", "delta"):
-                v = flat.get(k)
-                if isinstance(v, (dict, list)):
-                    flat[k] = json.dumps(v, ensure_ascii=False)[:500]
-            w.writerow(flat)
-    print(
-        f"wrote {args.out} matrix={matrix_path} blockers={blockers} "
-        f"accepted={accepted} stop_rule_met={blockers == 0}"
-    )
-    return 0 if blockers == 0 else 1
-
+from eplus_parity_compare import *  # noqa: F403
 
 if __name__ == "__main__":
+    from eplus_parity_compare import main
+
     raise SystemExit(main())
